requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
google-api-python-client==2.118.0
pydantic==2.6.1
pydantic-settings==2.1.0
loguru==0.7.2
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
apscheduler==3.10.4
jinja2==3.1.3
pandas==2.2.0
matplotlib==3.8.3
plotly==5.18.0

# Testing
pytest==8.0.1
httpx==0.26.0
pytest-asyncio==0.23.5
